refactor(ui): drop API-key prop drilling in model picker components
ReasoningModelSelector and TranscriptionModelPicker each took 4–6 API
keys + setters as props from every parent. The keys are global (one
Zustand source of truth, identical at every call site), so threading
them through props produced pointless coupling and re-renders.

Both components now read API keys directly via atomic useSettingsStore
selectors. Parents (InferenceConfigEditor, NotesOnboarding,
SettingsPage, OnboardingFlow, MeetingSettings, UploadAudioView) shed
their pass-through wirings and orphaned useSettings destructures.

Parent-controlled props (current scope's model, provider, mode,
custom-per-scope URL/key) stay as props — those legitimately differ
per call site.

Mirrors the pattern Gabriel already applied to EnterpriseProviderConfig.
Per TkDodo / Zustand idiom: atomic selectors per field inside the
consumer, no useShallow, no custom hook — the architectural fix is
removing the drilling, not relocating it.
