+++ src/components/OnboardingFlow.tsx
@@ -504,12 +504,6 @@ export default function OnboardingFlow({ onComplete }: OnboardingFlowProps) {
                   ...(!isLocal && !isSignedIn ? { cloudTranscriptionMode: "byok" } : {}),
                 });
               }}
-              openaiApiKey={openaiApiKey}
-              setOpenaiApiKey={setOpenaiApiKey}
-              groqApiKey={groqApiKey}
-              setGroqApiKey={setGroqApiKey}
-              mistralApiKey={mistralApiKey}
-              setMistralApiKey={setMistralApiKey}
               customTranscriptionApiKey={customTranscriptionApiKey}
               setCustomTranscriptionApiKey={setCustomTranscriptionApiKey}
               cloudTranscriptionBaseUrl={cloudTranscriptionBaseUrl}
